requirements.txt
seaborn
altair
plotly
hvplot
openpyxl
pygobject
ipython
holoviews
geopandas
geoviews
panel
